requirements.txt
cryptography
argon2-cffi
supabase
